requirements.txt
streamlit==1.27.1
pandas==1.5.3
streamlit_lottie==0.0.5
plotly_express==0.4.0
openpyxl==3.0.10
xlrd==1.0.0